+++ src/terms/TermCombobox.vue
@@ -1,13 +1,15 @@
 <template>
   <div v-on-click-outside="blur">
-    <div
+    <drop
       class="p-2 flex rounded-t shadow-inner items-center"
       :class="{
         'rounded-b': !suggestions.length,
         incomplete: input,
         'bg-smoke-200 hover:bg-smoke-300': !terms.length,
         'bg-blue-100': terms.length,
+        'animate-throb': isDraggingTerm,
       }"
+      @drop="dropTerm"
     >
       <div class="flex-1 flex flex-wrap items-baseline gap-1">
         <Term
@@ -15,9 +17,12 @@
           :key="term.name"
           class="term-added text-md"
           :data="term"
-          :options="['goto']"
+          :options="[removeOption, goto]"
+          :secondary="secondary"
+          :draggable="true"
         >
           {{ term._label }}
+          <template v-if="secondary">– perifert</template>
           <icon
             icon="times"
             size="xs"
@@ -52,7 +57,7 @@
         :value="showHelp"
         :toggle="toggleHelp"
       />
-    </div>
+    </drop>
 
     <div v-show="suggestions.length" class="h-0 relative z-20">
       <CloseButton @click="setSuggestions([])" />
@@ -79,21 +84,28 @@
 </template>
 
 <script setup>
-import { ref } from "@vue/reactivity";
+import { computed, ref } from "vue";
+import { useStore } from "vuex";
 import { useToggle } from "@vueuse/core";
 import { vOnClickOutside } from "@vueuse/components";
+import debounce from "lodash/debounce";
+import { searchTerms } from "@/services/terms.service";
 import Term from "@/terms/Term.vue";
 import CloseButton from "@/components/CloseButton.vue";
-import { searchTerms } from "@/services/terms.service";
-import debounce from "lodash/debounce";
 import ToggleIcon from "@/components/ToggleIcon.vue";
 import InputHelp from "@/components/InputHelp.vue";
+import useTerms from "./terms.composable";
+import useTermOptions from "./termOptions.composable";
 
-const props = defineProps(["terms", "input-id", "help"]);
+const props = defineProps(["terms", "input-id", "help", "secondary"]);
 const emit = defineEmits(["add", "remove"]);
+const { commit, state } = useStore();
 const [showHelp, toggleHelp] = useToggle();
+const { remove: removeTerm, removeSecondary } = useTerms();
+const { goto } = useTermOptions();
 const input = ref("");
 const suggestions = ref([]);
+const isDraggingTerm = computed(() => state.dragged?.type == "term");
 
 // No need to load suggestions until user slows down typing.
 const getSuggestions = debounce(async () => {
@@ -133,6 +145,24 @@ function setSuggestions(matches) {
 
 function blur() {
   setSuggestions([]);
+}
+
+const removeOption = (term) => ({
+  label: `Rensa <em>${term._label}</em>`,
+  action: () => remove(term),
+});
+
+function dropTerm() {
+  if (isDraggingTerm.value) {
+    const term = state.dragged.data;
+    // Remove from both fields.
+    removeTerm(term);
+    removeSecondary(term);
+    // Add to the current field.
+    add(term);
+
+    commit("setDragged", null);
+  }
 }
 </script>